pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/main_app"]

[project]
name = "personal-website"
version = "0.1.0"
description = "Personal blog and portfolio website built with FastHTML"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml",
    "python-frontmatter",
    "markdown",
    "pygments",
    "uvicorn[standard]",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0", 
    "ruff>=0.12.8",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
# Ignore star import warnings for FastHTML projects
ignore = [
    "F403",  # Star imports are used; unable to detect undefined names
    "F405",  # Name may be undefined, or defined from star imports
]

# Allow module-level imports not at top of file for route registration
[tool.ruff.lint.per-file-ignores]
"src/main_app/app.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

